@@ -1,8 +1,8 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v3.12.4
-// source: mock.proto
+// 	protoc-gen-go v1.36.10
+// 	protoc        v6.33.0
+// source: mock/mock.proto
 
 package mock
 
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -21,20 +22,17 @@ const (
 )
 
 type Request struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Val           string                 `protobuf:"bytes,1,opt,name=val" json:"val,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Val string `protobuf:"bytes,1,opt,name=val,proto3" json:"val,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Request) Reset() {
 	*x = Request{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_mock_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_mock_mock_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Request) String() string {
@@ -44,8 +42,8 @@ func (x *Request) String() string {
 func (*Request) ProtoMessage() {}
 
 func (x *Request) ProtoReflect() protoreflect.Message {
-	mi := &file_mock_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	mi := &file_mock_mock_proto_msgTypes[0]
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -57,7 +55,7 @@ func (x *Request) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Request.ProtoReflect.Descriptor instead.
 func (*Request) Descriptor() ([]byte, []int) {
-	return file_mock_proto_rawDescGZIP(), []int{0}
+	return file_mock_mock_proto_rawDescGZIP(), []int{0}
 }
 
 func (x *Request) GetVal() string {
@@ -68,20 +66,17 @@ func (x *Request) GetVal() string {
 }
 
 type Response struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Val           string                 `protobuf:"bytes,1,opt,name=val" json:"val,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Val string `protobuf:"bytes,1,opt,name=val,proto3" json:"val,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Response) Reset() {
 	*x = Response{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_mock_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_mock_mock_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Response) String() string {
@@ -91,8 +86,8 @@ func (x *Response) String() string {
 func (*Response) ProtoMessage() {}
 
 func (x *Response) ProtoReflect() protoreflect.Message {
-	mi := &file_mock_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	mi := &file_mock_mock_proto_msgTypes[1]
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -104,7 +99,7 @@ func (x *Response) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Response.ProtoReflect.Descriptor instead.
 func (*Response) Descriptor() ([]byte, []int) {
-	return file_mock_proto_rawDescGZIP(), []int{1}
+	return file_mock_mock_proto_rawDescGZIP(), []int{1}
 }
 
 func (x *Response) GetVal() string {
@@ -114,91 +109,65 @@ func (x *Response) GetVal() string {
 	return ""
 }
 
-var File_mock_proto protoreflect.FileDescriptor
+var File_mock_mock_proto protoreflect.FileDescriptor
 
-var file_mock_proto_rawDesc = []byte{
-	0x0a, 0x0a, 0x6d, 0x6f, 0x63, 0x6b, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x04, 0x6d, 0x6f,
-	0x63, 0x6b, 0x22, 0x1b, 0x0a, 0x07, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x10, 0x0a,
-	0x03, 0x76, 0x61, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x76, 0x61, 0x6c, 0x22,
-	0x1c, 0x0a, 0x08, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x76,
-	0x61, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x76, 0x61, 0x6c, 0x42, 0x24, 0x5a,
-	0x22, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x72, 0x65, 0x6c, 0x61,
-	0x62, 0x2f, 0x67, 0x6f, 0x72, 0x75, 0x6d, 0x73, 0x2f, 0x74, 0x65, 0x73, 0x74, 0x73, 0x2f, 0x6d,
-	0x6f, 0x63, 0x6b, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_mock_mock_proto_rawDesc = "" +
+	"\n" +
+	"\x0fmock/mock.proto\x12\x04mock\"\x1b\n" +
+	"\aRequest\x12\x10\n" +
+	"\x03val\x18\x01 \x01(\tR\x03val\"\x1c\n" +
+	"\bResponse\x12\x10\n" +
+	"\x03val\x18\x01 \x01(\tR\x03val2/\n" +
+	"\x06Server\x12%\n" +
+	"\x04Test\x12\r.mock.Request\x1a\x0e.mock.ResponseB)Z\"github.com/relab/gorums/tests/mock\x92\x03\x02\b\x02b\beditionsp\xe8\a"
 
 var (
-	file_mock_proto_rawDescOnce sync.Once
-	file_mock_proto_rawDescData = file_mock_proto_rawDesc
+	file_mock_mock_proto_rawDescOnce sync.Once
+	file_mock_mock_proto_rawDescData []byte
 )
 
-func file_mock_proto_rawDescGZIP() []byte {
-	file_mock_proto_rawDescOnce.Do(func() {
-		file_mock_proto_rawDescData = protoimpl.X.CompressGZIP(file_mock_proto_rawDescData)
+func file_mock_mock_proto_rawDescGZIP() []byte {
+	file_mock_mock_proto_rawDescOnce.Do(func() {
+		file_mock_mock_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_mock_mock_proto_rawDesc), len(file_mock_mock_proto_rawDesc)))
 	})
-	return file_mock_proto_rawDescData
+	return file_mock_mock_proto_rawDescData
 }
 
-var file_mock_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_mock_proto_goTypes = []interface{}{
+var file_mock_mock_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
+var file_mock_mock_proto_goTypes = []any{
 	(*Request)(nil),  // 0: mock.Request
 	(*Response)(nil), // 1: mock.Response
 }
-var file_mock_proto_depIdxs = []int32{
-	0, // [0:0] is the sub-list for method output_type
-	0, // [0:0] is the sub-list for method input_type
+var file_mock_mock_proto_depIdxs = []int32{
+	0, // 0: mock.Server.Test:input_type -> mock.Request
+	1, // 1: mock.Server.Test:output_type -> mock.Response
+	1, // [1:2] is the sub-list for method output_type
+	0, // [0:1] is the sub-list for method input_type
 	0, // [0:0] is the sub-list for extension type_name
 	0, // [0:0] is the sub-list for extension extendee
 	0, // [0:0] is the sub-list for field type_name
 }
 
-func init() { file_mock_proto_init() }
-func file_mock_proto_init() {
-	if File_mock_proto != nil {
+func init() { file_mock_mock_proto_init() }
+func file_mock_mock_proto_init() {
+	if File_mock_mock_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_mock_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Request); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_mock_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Response); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_mock_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_mock_mock_proto_rawDesc), len(file_mock_mock_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
-			NumServices:   0,
+			NumServices:   1,
 		},
-		GoTypes:           file_mock_proto_goTypes,
-		DependencyIndexes: file_mock_proto_depIdxs,
-		MessageInfos:      file_mock_proto_msgTypes,
+		GoTypes:           file_mock_mock_proto_goTypes,
+		DependencyIndexes: file_mock_mock_proto_depIdxs,
+		MessageInfos:      file_mock_mock_proto_msgTypes,
 	}.Build()
-	File_mock_proto = out.File
-	file_mock_proto_rawDesc = nil
-	file_mock_proto_goTypes = nil
-	file_mock_proto_depIdxs = nil
+	File_mock_mock_proto = out.File
+	file_mock_mock_proto_goTypes = nil
+	file_mock_mock_proto_depIdxs = nil
 }
