@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.9.0-devel+d62f9a0e
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+1fa869ba
+// 	protoc            v6.33.0
 // source: zorums.proto
 
 package dev
