@@ -26,29 +26,26 @@ type {{$service}}Server interface {
 var registerInterface = `
 {{$genFile := .GenFile}}
 {{$gorumsMessage := use "gorums.Message" .GenFile}}
-{{$wrapMessage := use "gorums.WrapMessage" $genFile}}
-{{$sendMessage := use "gorums.SendMessage" $genFile}}
+{{$newMessage := use "gorums.NewResponseMessage" $genFile}}
 {{range .Services -}}
 {{$service := .GoName}}
 func Register{{$service}}Server(srv *{{use "gorums.Server" $genFile}}, impl {{$service}}Server) {
 	{{- range .Methods}}
-	srv.RegisterHandler("{{.Desc.FullName}}", func(ctx {{$context}}, in *{{$gorumsMessage}}, {{if isOneway .}} _ {{- else}} finished {{- end}} chan<- *{{$gorumsMessage}}) {
+	srv.RegisterHandler("{{.Desc.FullName}}", func(ctx {{$context}}, in *{{$gorumsMessage}}) (*{{$gorumsMessage}}, error) {
 		req := in.Message.(*{{in $genFile .}})
-		defer ctx.Release()
 		{{- if isOneway .}}
 		impl.{{.GoName}}(ctx, req)
+		return nil, nil
 		{{- else if correctableStream .}}
 		err := impl.{{.GoName}}(ctx, req, func(resp *{{out $genFile .}}) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := {{use "proto.Clone" $genFile}}(in.Metadata)
-			return {{$sendMessage}}(ctx, finished, {{$wrapMessage}}(md.(*{{use "ordering.Metadata" $genFile}}), resp, nil))
+			// create a copy of the metadata, to avoid a data race between NewResponseMessage and SendMsg
+			md := {{use "proto.CloneOf" $genFile}}(in.Metadata)
+			return ctx.SendMessage({{$newMessage}}(md, resp))
 		})
-		if err != nil {
-			{{$sendMessage}}(ctx, finished, {{$wrapMessage}}(in.Metadata, nil, err))
-		}
+		return nil, err
 		{{- else }}
 		resp, err := impl.{{.GoName}}(ctx, req)
-		{{$sendMessage}}(ctx, finished, {{$wrapMessage}}(in.Metadata, resp, err))
+		return {{$newMessage}}(in.Metadata, resp), err
 		{{- end}}
 	})
 	{{- end}}
