@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.9.0-devel+ee38614a
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+e9eeea95
+// 	protoc            v6.33.0
 // source: config/config.proto
 
 package config
@@ -197,11 +197,10 @@ type ConfigTestServer interface {
 }
 
 func RegisterConfigTestServer(srv *gorums.Server, impl ConfigTestServer) {
-	srv.RegisterHandler("config.ConfigTest.Config", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+	srv.RegisterHandler("config.ConfigTest.Config", func(ctx gorums.ServerCtx, in *gorums.Message) (*gorums.Message, error) {
 		req := in.Message.(*Request)
-		defer ctx.Release()
 		resp, err := impl.Config(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
+		return gorums.NewResponseMessage(in.Metadata, resp), err
 	})
 }
 
