@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.9.0-devel+ee38614a
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+e9eeea95
+// 	protoc            v6.33.0
 // source: correctable/correctable.proto
 
 package correctable
@@ -10,7 +10,6 @@ import (
 	context "context"
 	fmt "fmt"
 	gorums "github.com/relab/gorums"
-	ordering "github.com/relab/gorums/ordering"
 	encoding "google.golang.org/grpc/encoding"
 	proto "google.golang.org/protobuf/proto"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
@@ -229,23 +228,19 @@ type CorrectableTestServer interface {
 }
 
 func RegisterCorrectableTestServer(srv *gorums.Server, impl CorrectableTestServer) {
-	srv.RegisterHandler("correctable.CorrectableTest.Correctable", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+	srv.RegisterHandler("correctable.CorrectableTest.Correctable", func(ctx gorums.ServerCtx, in *gorums.Message) (*gorums.Message, error) {
 		req := in.Message.(*CorrectableRequest)
-		defer ctx.Release()
 		resp, err := impl.Correctable(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
+		return gorums.NewResponseMessage(in.Metadata, resp), err
 	})
-	srv.RegisterHandler("correctable.CorrectableTest.CorrectableStream", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+	srv.RegisterHandler("correctable.CorrectableTest.CorrectableStream", func(ctx gorums.ServerCtx, in *gorums.Message) (*gorums.Message, error) {
 		req := in.Message.(*CorrectableRequest)
-		defer ctx.Release()
 		err := impl.CorrectableStream(ctx, req, func(resp *CorrectableResponse) error {
-			// create a copy of the metadata, to avoid a data race between WrapMessage and SendMsg
-			md := proto.Clone(in.Metadata)
-			return gorums.SendMessage(ctx, finished, gorums.WrapMessage(md.(*ordering.Metadata), resp, nil))
+			// create a copy of the metadata, to avoid a data race between NewResponseMessage and SendMsg
+			md := proto.CloneOf(in.Metadata)
+			return ctx.SendMessage(gorums.NewResponseMessage(md, resp))
 		})
-		if err != nil {
-			gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, nil, err))
-		}
+		return nil, err
 	})
 }
 
