@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-gorums. DO NOT EDIT.
 // versions:
-// 	protoc-gen-gorums v0.9.0-devel+ee38614a
-// 	protoc            v5.29.3
+// 	protoc-gen-gorums v0.9.0-devel+e9eeea95
+// 	protoc            v6.33.0
 // source: dummy/dummy.proto
 
 package dummy
@@ -180,10 +180,9 @@ type DummyServer interface {
 }
 
 func RegisterDummyServer(srv *gorums.Server, impl DummyServer) {
-	srv.RegisterHandler("dummy.Dummy.Test", func(ctx gorums.ServerCtx, in *gorums.Message, finished chan<- *gorums.Message) {
+	srv.RegisterHandler("dummy.Dummy.Test", func(ctx gorums.ServerCtx, in *gorums.Message) (*gorums.Message, error) {
 		req := in.Message.(*Empty)
-		defer ctx.Release()
 		resp, err := impl.Test(ctx, req)
-		gorums.SendMessage(ctx, finished, gorums.WrapMessage(in.Metadata, resp, err))
+		return gorums.NewResponseMessage(in.Metadata, resp), err
 	})
 }
