@@ -2,8 +2,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.6
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.10
+// 	protoc        v6.33.0
 // source: qf/qf.proto
 
 package qf
