@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.6
-// 	protoc        v5.29.3
+// 	protoc-gen-go v1.36.10
+// 	protoc        v6.33.0
 // source: unresponsive/unresponsive.proto
 
 package unresponsive
